src/main/java/ru/job4j/collection/ConvertList2Array.java
package ru.job4j.collection;

import java.util.List;

public class ConvertList2Array {
    public static int[][] toArray(List<Integer> list, int cells) {
        int groups = (int) Math.ceil((double) list.size() / cells);
        int[][] array = new int[groups][cells];
        int row = 0, cell = 0;
        for (Integer num : list) {
            array[row][cell] = num;
            cell++;
            if (cell == cells) {
                if (row <= groups - 1) {
                    row++;
                    cell = 0;
                } else {
                    break;
                }
            }
        }
        return array;
    }

    public static void main(String[] args) {
        List<Integer> list = List.of(1, 2, 3, 4, 5, 6, 7);
        int[][] result = toArray(list, 3);
        for (int[] row: result) {
            for (int cell: row) {
                System.out.println(cell + " ");
            }
            System.out.println();
        }
    }
}
